+++ src/gui.js
@@ -87,7 +87,7 @@ HatBlockMorph, ZOOM*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.gui = '2026-March-10';
+modules.gui = '2026-April-02';
 
 // Declarations
 
@@ -5921,12 +5921,15 @@ IDE_Morph.prototype.aboutSnap = function () {
         + '\nAchal Dave: Web Audio'
         + '\nJoe Otto: Morphic Testing and Debugging'
         + '\n\n'
+        + 'Ketrina Johnson created every Snap! costume and background'
+        + '\nexcept the ones inherited from Scratch, Alonzo, and the following:'
+        + '\n\n'
         + 'Jahrd, Derec, Jamet, Sarron, Aleassa, and Lirin costumes'
-        + '\nare watercolor paintings by Meghan Taylor and represent'
-        + '\n characters from her webcomic Prophecy of the Circle,'
-        + '\nlicensed to us only for use in Snap! projects.'
-        + '\nMeghan also painted the Tad costumes,'
-        + '\nbut that character is in the public domain.';
+        + 'are watercolor paintings\nby Meghan Taylor and represent'
+        + 'characters from her webcomic\nProphecy of the Circle,'
+        + 'licensed to us only for use in Snap! projects. '
+        + '\nMeghan also painted the Tad costumes, '
+        + 'but that character is in the public domain.';
 
     for (module in modules) {
         if (Object.prototype.hasOwnProperty.call(modules, module)) {