+++ HISTORY.md
@@ -68,7 +68,7 @@
     * new help screens
         * new help screens for "combinations" and
         * "pipe" reporters, thanks, Brian and gang!
-    * new set of costumes by Ketrina!
+    * new set of costumes and backgrounds by Ketrina!
 * **Notable Changes:**
     * data types
         * enhanced "is ... a ...?" predicate to support dependent data types and user defined structs (ADTs)
@@ -119,6 +119,7 @@
 ### 2026-04-02
 * threads: new feature: using a ring as index inside the "ITEM ... OF ..." reporter answers a copy of the function (ring) that is bound to the list object in the second slot, enabling polymorphic methods, i.e. message dispatches to a "super class"
 * updated shapes library with ability to rotate regular polygons
+* added Ketrina to the credits
 
 ### 2026-04-01
 * updated shapes library